464: Moved useMediaQuery and useWindowSize to hooks folder
