+++ src/components/Page/Page.tsx
@@ -8,9 +8,9 @@ import { useWeb3React } from "@web3-react/core";
 import { useAppDispatch } from "../../app/hooks";
 import { resetOrders } from "../../features/orders/ordersSlice";
 import { Wallet } from "../../features/wallet/Wallet";
-import useWindowSize from "../../helpers/useWindowSize";
 import useAppRouteParams from "../../hooks/useAppRouteParams";
 import useDebounce from "../../hooks/useDebounce";
+import useWindowSize from "../../hooks/useWindowSize";
 import { AppRoutes } from "../../routes";
 import HelmetContainer from "../HelmetContainer/HelmetContainer";
 import { InformationModalType } from "../InformationModals/InformationModals";
